[Chore #161361356] Restructure project structure
