+++ run.py
@@ -1,4 +1,4 @@
-from api_v1.app.app import app
+from app import app
 
 if __name__ == '__main__':
     app.run(debug=True, port=5001)
